requirements.txt
pyyaml~=6.0
pdfplumber~=0.10.4
tqdm~=4.65.0
openai~=1.10.0
jinja2~=3.1.2